gemfile: add basic Gemfile parser
Add basic lepl based parser for Gemfile. It supports only a
limit subset of ruby gramma, normally used in Gemfiles.

It parses successfully the xikolo-account Gemfile.

It is WIP, the parse should be combined with the Gemfile.lock parser
to return all needed information (e.g. also in which group a
gem is loaded).