rubygem: workaround for execuatable name detection
Gem like `factory_girl_rails` provide executable like `bin/stubs/rake`.
Currently we used basename to extract the name for invocation. This
resulted in name conflicts with the real gem (like `rake`).

For now we strip the first component (as it is the bin directory):
`bin/stubs/rake` -> `stubs/rake`. Later on we must store also the
bin_path to be able to calculate the correct relative name.

ref #3